mdns: NX_MDNS_GET_ULONG_DATA shifts a signed int
Reading the A record of any host whose address starts with a byte >= 128
is undefined behaviour. UBSan, on a datagram announcing 169.254.x.x:

  nxd_mdns.c:9611: runtime error: left shift of 169 by 24 places cannot
  be represented in type 'int'

The macro is

  #define NX_MDNS_GET_ULONG_DATA(data) \
      (ULONG)((*(data) << 24) | (*(data + 1) << 16) | \
              (*(data + 2) << 8) | (*(data + 3)))

`*(data)` is a UCHAR, so it integer-promotes to int, and `int << 24` for
any value >= 128 exceeds INT_MAX. C11 6.5.7p4 makes that undefined. The
(ULONG) cast is applied to the completed expression and so comes too
late to help.

This is not an edge case reachable only by hostile input: the first
octet is >= 128 for every address in 128.0.0.0/1, which includes
192.168.0.0/16 and 169.254.0.0/16. Ordinary mDNS traffic on an ordinary
network reaches it, and _nx_mdns_process_resource_record() calls it for
every A record parsed.

It works today because GCC and Clang lower the shift to the wrapping
instruction the author intended. It is still undefined, it fails a
UBSan build, and a compiler entitled to assume no signed overflow is
entitled to do something else with it.

Cast each byte to ULONG before shifting, which is what the trailing
cast was reaching for. No behaviour change on any compiler that was
already doing the expected thing.

Found by fuzzing the mDNS response parser under ASan and UBSan with
malformed and well-formed datagrams; this fires on the well-formed ones.

The sibling accessor in addons/dns is not affected -- it does not use
this pattern.

+++ addons/mdns/nxd_mdns.h
@@ -67,7 +67,7 @@ extern   "C" {
 
 /* Define macros that get the data on unaligned address.  */
 #define NX_MDNS_GET_USHORT_DATA(data)           (USHORT)((*(data) << 8) | (*(data + 1)))
-#define NX_MDNS_GET_ULONG_DATA(data)            (ULONG)((*(data) << 24) | (*(data + 1) << 16) | (*(data + 2) << 8) | (*(data + 3)))
+#define NX_MDNS_GET_ULONG_DATA(data)            (((ULONG)(*(data)) << 24) | ((ULONG)(*(data + 1)) << 16) | ((ULONG)(*(data + 2)) << 8) | ((ULONG)(*(data + 3))))
 
 /* Disable the mDNS Server functionality.  By default mDNS server
    function is enabled.  To remove the mDNS server function from
